+++ test/registry_test.exs
@@ -57,7 +57,7 @@ defmodule RegistryTest do
         Horde.Registry.start_link(
           name: :horde_not_unique,
           keys: :duplicate,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
       end
     end
@@ -339,14 +339,14 @@ defmodule RegistryTest do
         Horde.Registry.start_link(
           name: horde1,
           keys: :unique,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       {:ok, _} =
         Horde.Registry.start_link(
           name: horde2,
           keys: :unique,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       Horde.Cluster.set_members(horde1, [horde1, horde2])
@@ -513,7 +513,7 @@ defmodule RegistryTest do
         Horde.Registry.start_link(
           name: :match_horde,
           keys: :unique,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       Horde.Registry.register(:match_horde, "foo", "foo")
@@ -542,7 +542,7 @@ defmodule RegistryTest do
         Horde.Registry.start_link(
           name: :update_value_horde,
           keys: :unique,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       Horde.Registry.register(:update_value_horde, "foo", "foo")
@@ -562,7 +562,7 @@ defmodule RegistryTest do
         Horde.Registry.start_link(
           name: :update_value_horde2,
           keys: :unique,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       Task.start_link(fn ->
@@ -585,7 +585,7 @@ defmodule RegistryTest do
         Horde.Registry.start_link(
           name: horde,
           keys: :unique,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       Horde.Registry.register(horde, :carmen, "foo")
@@ -600,7 +600,7 @@ defmodule RegistryTest do
         Horde.Registry.start_link(
           name: horde,
           keys: :unique,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       Horde.Registry.register(horde, :carmen, "bar")
@@ -618,7 +618,7 @@ defmodule RegistryTest do
         Horde.Registry.start_link(
           name: horde,
           keys: :unique,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       Horde.Registry.register(horde, :carmen, "carmen")
@@ -635,7 +635,7 @@ defmodule RegistryTest do
         Horde.Registry.start_link(
           name: horde,
           keys: :unique,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       spawn(fn ->
@@ -657,7 +657,7 @@ defmodule RegistryTest do
         Horde.Registry.start_link(
           name: horde,
           keys: :unique,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       spawn(fn ->
@@ -680,7 +680,7 @@ defmodule RegistryTest do
           name: registry,
           keys: :unique,
           meta: [meta_key: :meta_value],
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       assert {:ok, :meta_value} = Horde.Registry.meta(registry, :meta_key)
@@ -693,7 +693,7 @@ defmodule RegistryTest do
         Horde.Registry.start_link(
           name: registry,
           keys: :unique,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       :ok = Horde.Registry.put_meta(registry, :custom_key, "custom_value")
@@ -708,7 +708,7 @@ defmodule RegistryTest do
         Horde.Registry.start_link(
           name: registry,
           keys: :unique,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       :ok = Horde.Registry.put_meta(registry, :custom_key, "custom_value")
@@ -724,14 +724,14 @@ defmodule RegistryTest do
         Horde.Registry.start_link(
           name: r1,
           keys: :unique,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       {:ok, _horde} =
         Horde.Registry.start_link(
           name: r2,
           keys: :unique,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       Horde.Cluster.set_members(r1, [r1, r2])
@@ -749,7 +749,7 @@ defmodule RegistryTest do
         Horde.Registry.start_link(
           name: registry,
           keys: :unique,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       %{pid: pid} =
@@ -777,14 +777,14 @@ defmodule RegistryTest do
           name: reg1,
           keys: :unique,
           members: [reg1, reg2],
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       {:ok, _} =
         Horde.Registry.start_link(
           name: reg2,
           keys: :unique,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       %{pid: pid} =
@@ -888,7 +888,7 @@ defmodule RegistryTest do
     horde = :"h#{-:erlang.monotonic_time()}"
 
     {:ok, _pid} =
-      Horde.Registry.start_link([name: horde, delta_crdt_options: [sync_interval: 20]] ++ opts)
+      Horde.Registry.start_link([name: horde, delta_crdt: [sync_interval: 20]] ++ opts)
 
     horde
   end