+++ test/supervisor_test.exs
@@ -11,21 +11,21 @@ defmodule SupervisorTest do
       Horde.Supervisor.start_link(
         name: n1,
         strategy: :one_for_one,
-        delta_crdt_options: [sync_interval: 20]
+        delta_crdt: [sync_interval: 20]
       )
 
     {:ok, _} =
       Horde.Supervisor.start_link(
         name: n2,
         strategy: :one_for_one,
-        delta_crdt_options: [sync_interval: 20]
+        delta_crdt: [sync_interval: 20]
       )
 
     {:ok, _} =
       Horde.Supervisor.start_link(
         name: n3,
         strategy: :one_for_one,
-        delta_crdt_options: [sync_interval: 20]
+        delta_crdt: [sync_interval: 20]
       )
 
     Horde.Cluster.set_members(n1, [n1, n2, n3])
@@ -84,8 +84,8 @@ defmodule SupervisorTest do
 
   describe "module-based Supervisor" do
     test "can use `init` function to dynamically fetch configuration" do
-      {:ok, _} = TestSupervisor1.start_link(name: :init_sup_test_1, strategy: :one_for_one)
-      {:ok, _} = TestSupervisor2.start_link(name: :init_sup_test_2, strategy: :one_for_one)
+      {:ok, _} = TestSupervisor1.start_link([], name: :init_sup_test_1)
+      {:ok, _} = TestSupervisor2.start_link([], name: :init_sup_test_2)
       members = Horde.Cluster.members(:init_sup_test_1)
       assert 2 = Enum.count(members)
     end
@@ -282,14 +282,14 @@ defmodule SupervisorTest do
         Horde.Supervisor.start_link(
           name: :horde_1_graceful,
           strategy: :one_for_one,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       {:ok, _} =
         Horde.Supervisor.start_link(
           name: :horde_2_graceful,
           strategy: :one_for_one,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       defmodule TerminationDelay do
@@ -355,7 +355,7 @@ defmodule SupervisorTest do
         Horde.Supervisor.start_link(
           name: :horde_transient,
           strategy: :one_for_one,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       Horde.Supervisor.start_child(:horde_transient, child_spec)
@@ -374,7 +374,7 @@ defmodule SupervisorTest do
         Horde.Supervisor.start_link(
           name: :horde_transient,
           strategy: :one_for_one,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
       Horde.Supervisor.start_child(:horde_transient, child_spec)
@@ -393,10 +393,10 @@ defmodule SupervisorTest do
         Horde.Supervisor.start_link(
           name: :horde_transient,
           strategy: :one_for_one,
-          delta_crdt_options: [sync_interval: 20]
+          delta_crdt: [sync_interval: 20]
         )
 
-      Horde.Supervisor.start_child(:horde_transient, child_spec)
+        Horde.Supervisor.start_child(:horde_transient, child_spec)
 
       processes = :sys.get_state(:horde_transient).processes_by_id
       assert Enum.count(processes) == 1
@@ -447,7 +447,7 @@ defmodule SupervisorTest do
         strategy: :one_for_one,
         distribution_strategy: Horde.UniformQuorumDistribution,
         members: [:horde_quorum_1, :horde_quorum_2, :horde_quorum_3],
-        delta_crdt_options: [sync_interval: 20]
+        delta_crdt: [sync_interval: 20]
       )
 
     catch_exit(Horde.Supervisor.wait_for_quorum(:horde_quorum_1, 100))
@@ -458,7 +458,7 @@ defmodule SupervisorTest do
         strategy: :one_for_one,
         distribution_strategy: Horde.UniformQuorumDistribution,
         members: [:horde_quorum_1, :horde_quorum_2, :horde_quorum_3],
-        delta_crdt_options: [sync_interval: 20]
+        delta_crdt: [sync_interval: 20]
       )
 
     assert :ok == Horde.Supervisor.wait_for_quorum(:horde_quorum_1, 1000)